requirements.txt
streamlit
mysql-connector-python
faker
pandas
matplotlib
